Delete old files, rename new impl
