+++ src/definitions/Character.ts
@@ -1,76 +1,6 @@
-// // import deepmerge from "deepmerge";
-// import { RefType, t, td } from "../Util";
-// import { computed, ComputedRef, isReadonly, isRef, reactive, Ref, ref, toRaw, toRefs, unref, watch, WritableComputedRef } from "vue";
-// import Merit, { MERITS } from "./Merit";
-// import { EnumSplat, Form, FormMods, Organization, Splat, SPLATS, SubType, WEREWOLF_FORMS } from "./Splat";
-// import { key } from "localforage";
-
-// export const ATTRIBUTES = [
-// 	["intelligence", "wits", "resolve"],
-// 	["strength", "dexterity", "stamina"],
-// 	["presence", "manipulation", "composure"],
-// ];
-
-// export const SKILLS = [
-// 	[
-// 		"academics",
-// 		"computer",
-// 		"crafts",
-// 		"investigation",
-// 		"medicine",
-// 		"occult",
-// 		"politics",
-// 		"science",
-// 	],
-// 	[
-// 		"athletics",
-// 		"brawl",
-// 		"drive",
-// 		"firearms",
-// 		"larceny",
-// 		"stealth",
-// 		"survival",
-// 		"weaponry",
-// 	],
-// 	[
-// 		"animal_ken",
-// 		"empathy",
-// 		"expression",
-// 		"intimidation",
-// 		"persuasion",
-// 		"socialize",
-// 		"streetwise",
-// 		"subterfuge",
-// 	],
-// ];
-
-// export interface Ability {
-// 	key?: string;
-// 	name: RefType<string>;
-// 	level: number;
-// }
-
-
-// function sortObj(obj: any) {
-// 	return Object.keys(obj).sort().reduce((result, key) => {
-// 		(result as any)[key] = obj[key];
-// 		return result;
-// 	}, {});
-// }
-
-// export function getNum(val: string): number {
-// 	// return computed(() => {
-// 	// const ret = (() => {
-// 	try {
-// 		return eval(val) || 0;
-// 	} catch (e) {
-// 		return 0;
-// 	}
-// 	// })();
-
-// 	// return isRef(ret) ? ret : ref(ret);
-// 	// });
-// }
+import _ from "lodash";
+import { reactive } from "vue";
+import { Splat, SPLATS, EnumSplat, Organization, SubType, MERITS, Merit, Modifier } from ".";
 
 export function nameToKey(name: string) {
 	return name.trim().toLowerCase().replaceAll(" ", "_").replaceAll("-", "_");//.replaceAll("(", "_").replaceAll(")", "_");
@@ -80,846 +10,883 @@ export function keyToName(key: string) {
 	return key.replaceAll("_", " ");
 }
 
-// export function def<T>(func: () => T, def?: T): ComputedRef<T> {
-// 	if (!def) def = 0 as any;
-// 	return computed(() => {
-// 		try {
-// 			const val = func();
-// 			// console.log(def);
-// 			// if (def === undefined) {
-// 			// 	if (typeof val === "number") {
-// 			// 		def = 0 as any;
-// 			// 	}
-// 			// }
-// 			// console.log(val, typeof val);
 
-// 			return val || def;
-// 		} catch (e) {
-// 			return def || 0;
-// 		}
-// 	}) as any;
-// 	// try {
-// 	// 	return func() || def;
-// 	// } catch (e) {
-// 	// 	return def;
-// 	// }
+export const ATTRIBUTES = [
+	["intelligence", "wits", "resolve"],
+	["strength", "dexterity", "stamina"],
+	["presence", "manipulation", "composure"],
+];
+
+export const SKILLS = [
+	[
+		"academics",
+		"computer",
+		"crafts",
+		"investigation",
+		"medicine",
+		"occult",
+		"politics",
+		"science",
+	],
+	[
+		"athletics",
+		"brawl",
+		"drive",
+		"firearms",
+		"larceny",
+		"stealth",
+		"survival",
+		"weaponry",
+	],
+	[
+		"animal_ken",
+		"empathy",
+		"expression",
+		"intimidation",
+		"persuasion",
+		"socialize",
+		"streetwise",
+		"subterfuge",
+	],
+];
+export interface Ability {
+	key?: string;
+	name?: string;
+	level: number;
+}
+
+export interface Attributes {
+	[index: string]: number;
+	intelligence: number;
+	wits: number;
+	resolve: number;
+
+	strength: number;
+	dexterity: number;
+	stamina: number;
+
+	presence: number;
+	manipulation: number;
+	composure: number;
+}
+
+interface Skills {
+	[index: string]: number;
+}
+
+interface Armor {
+	[key: string]: number;
+	general: number;
+	ballistic: number;
+}
+
+interface ICharacter {
+	splat: Splat;
+	name: string;
+
+	concept: string;
+	chronicle: string;
+
+	organization: Organization;
+	subType: SubType;
+
+	// xsplat: string;
+	// ysplat: string;
+	// zsplat: string;
+
+	virtueAnchor: string;
+	viceAnchor: string;
+
+	attributes: Attributes;
+	skills: Skills;
+
+	maxHealth: number;
+	healthTrack: number[];
+	woundPenalty: number;
+
+	conditions: string[];
+	aspirations: string[];
+
+	size: number;
+	speed: number;
+	defense: number;
+	armor: Armor;
+	initative: number;
+	perception: number;
+
+	willpower: number;
+	maxWillpower: number;
+	spentWillpowerDots: number;
+
+	mod(string: string): number;
+
+	data: Map<string, unknown>;
+}
+
+export class Character implements ICharacter {
+
+	get splat(): Splat {
+		return SPLATS[this.data.get("splat") as EnumSplat] || {
+			name: "",
+			organizations: {},
+			subTypes: {}
+		};
+	}
+
+	set splat(val: Splat) {
+		this.data.set("splat", val.enum);
+	}
+
+	name: string;
+
+	concept: string;
+	chronicle: string;
+
+	virtueAnchor: string;
+	viceAnchor: string;
+
+	get organization(): Organization {
+		return this.splat.organizations[this.data.get("organization") as string] || Object.defineProperties({}, {
+			name: {
+				get: () => this.data.get("organization"),
+				set: (val: string) => this.data.set("organization", val)
+			}
+		});
+	}
+
+	get subType(): SubType {
+		return this.splat.subTypes[this.data.get("subType") as string] || {};
+	}
+
+	attributes: Attributes;
+	skills: Skills;
+
+	get maxHealth(): number {
+		return this.attributes.stamina + this.size;
+	}
+
+	get healthTrack(): number[] {
+		const ogTrack = (this.data.get("healthTrack") as number[]) || [];
+		const track = Object.assign(new Array(this.maxHealth || 0).fill(0), ogTrack);
+
+		if (this.maxHealth < track.length) {
+			const deleted = track.splice(this.maxHealth - 1, track.length - this.maxHealth);
+		}
+
+		if (JSON.stringify(track) !== JSON.stringify(ogTrack)) {
+			this.data.set("healthTrack", track);
+		}
+
+		return track;
+	}
+
+	set healthTrack(track: number[]) {
+		this.data.set("healthTrack", track);
+	}
+
+	get woundPenalty(): number {
+		return Math.min((this.healthTrack[this.maxHealth - 1] !== 0 ?
+			-3 : this.healthTrack[this.maxHealth - 2] !== 0 ?
+				-2 : this.healthTrack[this.maxHealth - 3] !== 0 ?
+					-1 : 0), 0);
+	}
+
+	conditions: string[] = [];
+	aspirations: string[] = [];
+
+	willpower = 0;
+
+	get maxWillpower(): number {
+		return this.attributes.resolve + this.attributes.composure - this.spentWillpowerDots;
+	}
+
+	get spentWillpowerDots(): number {
+		return this.data.get("spentWillpowerDots") as number;
+	}
+	set spentWillpowerDots(val: number) {
+		this.data.set("spentWillpowerDots", val);
+	}
+
+	// size: number;
+
+	get size(): number {
+		return this.data.get("size") as number;
+	}
+
+	set size(val: number) {
+		this.data.set("size", val);
+	}
+
+	get speed(): number {
+		return this.attributes.strength + this.attributes.dexterity + 5;
+	}
+
+	get defense(): number {
+		return Math.min(this.attributes.dexterity, this.attributes.wits) + this.skills.athletics;
+	}
+
+	get perception(): number {
+		return this.attributes.wits + this.attributes.composure;
+	}
+
+	armor: Armor;
+
+	// get armor(): Armor {
+	// 	const data = this.data.get("armor") as Armor || {
+	// 		general: 0,
+	// 		ballistic: 0
+	// 	};
+
+	// 	const obj = {} as Armor;
+
+	// 	Object.defineProperties(obj, {
+	// 		general: {
+	// 			get: () => data.general
+	// 		},
+	// 		ballistic: {
+	// 			get: () => data.ballistic
+	// 		}
+	// 	});
+
+	// 	return obj;
+	// }
+
+	get initative(): number {
+		return this.attributes.dexterity + this.attributes.composure;
+	}
+
+	// templates: EnumTemplate[];
+	data: Map<string, unknown> = reactive(new Map());
+
+	constructor(opts: ICharacter) {
+		if (typeof opts.splat === "number") {
+			this.data.set("splat", opts.splat);
+		} else if (typeof opts.splat === "object" && typeof opts.splat.enum === "number") {
+			this.data.set("splat", opts.splat.enum);
+		}
+
+		if (opts.subType) {
+			this.data.set("subType", opts.subType);
+		}
+
+		if (opts.organization) {
+			this.data.set("organization", opts.organization);
+		}
+
+		this.name = opts.name || "";
+
+		this.concept = opts.concept || "";
+		this.chronicle = opts.chronicle || "";
+
+		this.virtueAnchor = opts.virtueAnchor || "";
+		this.viceAnchor = opts.viceAnchor || "";
+
+		this.willpower = opts.willpower || 0;
+
+		this.data.set("attributes", Object.assign({
+			intelligence: 1,
+			wits: 1,
+			resolve: 1,
+			strength: 1,
+			dexterity: 1,
+			stamina: 1,
+			presence: 1,
+			manipulation: 1,
+			composure: 1
+		} as Attributes, opts.attributes));
+		this.data.set("skills", opts.skills || {} as Skills);
+
+		{
+			const _getAttrs = () => this.data.get("attributes") as Attributes;
+			const _setAttrs = (key: string, val: number) => this.data.set("attributes", Object.assign(_getAttrs(), { [key]: val }));
+			const _defProp = (key: string) => ({
+				get: () => (_getAttrs()[key] || 1) + this.mod(key),
+				set: (val: number) => _setAttrs(key, val - this.mod(key)),
+				configurable: true
+			});
+
+			this.attributes = Object.defineProperties({} as Attributes, {
+				intelligence: _defProp("intelligence"),
+				wits: _defProp("wits"),
+				resolve: _defProp("resolve"),
+
+				strength: _defProp("strength"),
+				dexterity: _defProp("dexterity"),
+				stamina: _defProp("stamina"),
+
+				presence: _defProp("presence"),
+				manipulation: _defProp("manipulation"),
+				composure: _defProp("composure"),
+			});
+		}
+
+		{
+			const _getSkills = () => this.data.get("skills") as Skills;
+			const _setSkill = (key: string, val: number) => this.data.set("skills", Object.assign(_getSkills(), { [key]: val }));
+
+			const props = {} as any;
+
+			SKILLS.flat().forEach(skill => {
+				props[skill] = {
+					get: () => _getSkills()[skill] || 0,
+					set: (val: number) => _setSkill(skill, val)
+				};
+			});
+
+			this.skills = Object.defineProperties({} as Skills, props);
+		}
+
+		this.data.set("armor", opts.armor || {});
+		{
+			const _getAttrs = () => this.data.get("armor") as Armor;
+			const _setAttrs = (key: string, val: number) => this.data.set("armor", Object.assign(_getAttrs(), { [key]: val }));
+			const _defProp = (key: string, func: () => number) => ({
+				get: () => (_getAttrs()[key] || 0) + func(),
+				set: (val: number) => _setAttrs(key, val - func())
+			});
+
+			this.armor = Object.defineProperties({} as Armor, {
+				general: _defProp("general", () => 0),
+				ballistic: _defProp("ballistic", () => 0),
+			});
+		}
+
+		// this.size = opts.size || 5;
+		this.data.set("size", opts.size || 5);
+
+		this.data.set("healthTrack", opts.healthTrack || []);
+
+		this.data.set("spentWillpowerDots", opts.spentWillpowerDots || 0);
+	}
+
+	getData(): any {
+		const clone = Object.assign({}, this);
+		this.data.forEach((val, key) => {
+			(clone as any)[key] = val;
+		});
+		return clone;
+	}
+
+	mod(trait: string): number {
+		return 0;
+	}
+}
+
+interface IMortalCharacter extends ICharacter {
+
+	specialties: { [index: string]: string[] };
+
+	merits: { [key: string]: Ability };
+
+	integrityTrait: number;
+
+	beats: number;
+	experience: number;
+}
+
+
+const MERIT_CACHE: {
+	[key: string]: {
+		[index: string]: Merit
+	}
+} = {};
+
+export class MortalCharacter extends Character implements IMortalCharacter {
+
+	specialties: { [index: string]: string[] };
+
+	beats: number;
+	experience: number;
+
+	// merits: { [key: string]: Ability };
+	get merits(): { [key: string]: Ability } {
+		const data = this.data.get("merits") as { [key: string]: Ability };
+
+		const obj = _.clone(data);
+
+		if (!MERIT_CACHE[this.id]) MERIT_CACHE[this.id] = {};
+
+		Object.entries(data).forEach(([key, value]) => {
+			const meritKey = key.includes("(") ? key.substr(0, key.indexOf("(") - 1) : key;
+
+			const m = MERITS[meritKey];
+			const cache = MERIT_CACHE[this.id][key];
+
+			if (cache) {
+				obj[key] = cache;
+			} else if (m) {
+				const merit = new m(this, value);
+				MERIT_CACHE[this.id][key] = merit;
+
+				obj[key] = merit;
+			}
+		});
+
+		return reactive(obj);
+	}
+
+	set merits(val: { [key: string]: Ability }) {
+		this.data.set("merits", val);
+	}
+
+	get meritTraitMods(): { [key: string]: number } {
+		const obj = Object.values(this.merits)
+			.filter(m => m instanceof Merit)
+			.flatMap((el) => (el as Merit).getTraitMods(this))
+			.map(el => {
+				return {
+					mod: typeof el.mod === "function" ? el.mod() : el.mod,
+					trait: el.trait
+				};
+			}).reduce((acc, val) => {
+				acc[val.trait] = val.mod + (acc[val.trait] || 0);
+
+				return acc;
+			}, {} as { [key: string]: number });
+
+		return obj;
+	}
+
+	get size(): number {
+		return super.size + (this.meritTraitMods.size || 0);
+	}
+
+	set size(val: number) {
+		super.size = val - (this.meritTraitMods.size || 0);
+	}
+
+	integrityTrait: number;
+
+	private id: string;
+	constructor(opts: IMortalCharacter) {
+		super(opts);
+		this.id = (opts as any).id || "";
+		this.data.set("merits", opts.merits || {});
+
+		this.beats = opts.beats || 0;
+		this.experience = opts.experience || 0;
+
+		this.specialties = opts.specialties || {};
+		this.integrityTrait = opts.integrityTrait || 7;
+	}
+
+}
+
+interface ISupernatural {
+
+	power: number;
+
+	abilities: { [index: string]: Ability };
+
+	maxFuel: number;
+	fuel: number;
+
+}
+
+interface ISupernaturalCharacter extends IMortalCharacter, ISupernatural {
+	power: number;
+	abilities: { [index: string]: Ability };
+	fuel: number;
+	maxFuel: number;
+}
+
+export class SupernaturalCharacter extends MortalCharacter implements ISupernaturalCharacter {
+
+	power: number;
+
+	abilities: { [index: string]: Ability } = {};
+
+	fuel: number;
+	get maxFuel(): number {
+		return this.power >= 5 ?
+			this.power >= 9 ?
+				this.power === 10 ?
+					75 :
+					50
+				: 10 + (this.power - 4) * 5
+			: 10 + this.power - 1;
+	}
+
+	constructor(opts: ISupernaturalCharacter) {
+		super(opts);
+
+		this.power = opts.power || 1;
+		this.fuel = opts.fuel || 0;
+
+		this.abilities = {};
+
+		{
+			const defaultAbl: { [index: string]: Ability } = {};
+			const custom: { [index: string]: Ability } = {};
+
+			Object.keys(this.splat.abilities || {}).forEach(key => {
+				defaultAbl[key] = { level: 0 };
+			});
+
+			const ablTemp = Object.assign({}, defaultAbl, opts.abilities);
+
+			Object.keys(ablTemp).forEach(key => {
+				const value = ablTemp[key];
+				const def = defaultAbl[key];
+
+				if (!def) {
+					custom[key] = value;
+					delete ablTemp[key];
+				} else {
+					// value.name = def.name;
+				}
+			});
+
+			Object.keys(custom)
+				.forEach(key => ablTemp[key] = custom[key]);
+			this.abilities = reactive(ablTemp);
+		}
+	}
+
+	protected _getAbility(key: string): Ability {
+		return this.abilities[key] || { level: 0, name: "" };
+	}
+}
+
+interface IMageCharacter extends IMortalCharacter, ISupernatural {
+
+	roteSkills: string[];
+
+	activeSpells: string[];
+	yantras: string[];
+	magicalTools: string[];
+	praxes: string[];
+	inuredSpells: string[];
+	nimbus: string[];
+	obsessions: string[];
+
+	attainments: string[];
+	legacyAttainments: string[];
+
+	rotes: Rote[];
+}
+
+
+export class Rote {
+	arcanum: string;
+	level: number;
+	spell: string;
+	creator?: string;
+	roteSkill: string;
+
+	constructor(opts: Rote) {
+		opts = opts || {};
+		this.arcanum = opts.arcanum || "";
+		this.level = opts.level || 0;
+
+		this.spell = opts.spell || "";
+		this.creator = opts.creator || "";
+
+		this.roteSkill = opts.roteSkill || "";
+	}
+}
+
+export class MageCharacter extends SupernaturalCharacter implements IMageCharacter {
+
+	// roteSkills: string[];
+
+	activeSpells: string[];
+	yantras: string[];
+	magicalTools: string[];
+	praxes: string[];
+	inuredSpells: string[];
+	nimbus: string[];
+	obsessions: string[];
+
+	attainments: string[];
+	legacyAttainments: string[];
+
+	rotes: Rote[];
+
+	get roteSkills(): string[] {
+		return this.organization.skills || this.data.get("roteSkills") as string[] || [];
+	}
+
+	constructor(opts: IMageCharacter) {
+		super(opts);
+
+		this.activeSpells = opts.activeSpells || [];
+		this.yantras = opts.yantras || [];
+		this.magicalTools = opts.magicalTools || [];
+		this.praxes = opts.praxes || [];
+		this.inuredSpells = opts.inuredSpells || [];
+		this.nimbus = opts.nimbus || [];
+
+		this.obsessions = opts.obsessions || [];
+
+		this.attainments = opts.attainments || [];
+		this.legacyAttainments = opts.legacyAttainments || [];
+
+
+		this.rotes = opts.rotes || [];
+
+		if (opts.roteSkills) {
+			this.data.set("roteSkills", opts.roteSkills);
+		}
+		// this.roteSkills = opts.roteSkills || [];
+	}
+
+}
+
+interface IHasTouchstones {
+
+	touchstones: string[];
+
+}
+
+interface IVampireCharacter extends IMortalCharacter, ISupernatural, IHasTouchstones {
+}
+
+function lookupAccess<T>(obj: any, prop: string) {
+	const desc = (Object.getOwnPropertyDescriptor(obj, prop) || {});
+
+	return {
+		get: desc.get || (() => null),
+		set: desc.set || ((val: T) => null)
+	} as {
+		get: () => T,
+		set: (val: T) => void
+	};
+}
+
+// function lookupAccessObj<T>(obj: any) {
+// 	Object.getOwnPropertyDescriptors(obj);
 // }
 
-// export interface Attributes {
-// 	[index: string]: number;
-// 	intelligence: number;
-// 	wits: number;
-// 	resolve: number;
+export class VampireCharacter extends SupernaturalCharacter implements IVampireCharacter {
 
-// 	strength: number;
-// 	dexterity: number;
-// 	stamina: number;
+	touchstones: string[];
 
-// 	presence: number;
-// 	manipulation: number;
-// 	composure: number;
-// }
+	get defense(): number {
+		return super.defense + this._getAbility("celerity").level;
+	}
 
-// // interface IntegrityTrack {}
+	get maxFuel(): number {
+		return this.power == 0 ? this.attributes.stamina : super.maxFuel;
+	}
 
-// export interface Armor {
-// 	general?: number;
-// 	ballistic?: number;
-// }
+	constructor(opts: IVampireCharacter) {
+		super(opts);
+		this.touchstones = opts.touchstones || [];
 
-// export interface TraitMod {
-// 	trait?: string;
-// 	mod?: Ref<number> | number;
-// 	func?: (traitName: string) => Ref<number> | number;
-// }
+		const stam = lookupAccess<number>(this.attributes, "stamina");
+		Object.defineProperty(this.attributes, "stamina", {
+			get: () => stam.get() + this._getAbility("resilience").level,
+		});
+		// set: (val: number) => stam.set(val - this._getAbility("resilience").level)
 
-// export interface Weapon {
-// 	name: string;
-// 	damage: string;
-// 	range: string;
-// 	clip: string;
-// 	initative: number;
-// 	strength: number;
-// 	size: number;
-// }
+		const str = lookupAccess<number>(this.attributes, "strength");
+		Object.defineProperty(this.attributes, "strength", {
+			get: () => str.get() + this._getAbility("vigor").level,
+		});
+	}
+}
 
-// export default class Character {
-// 	name!: string;
-// 	age?: number;
+interface IWerewolfCharacter extends IMortalCharacter, ISupernatural, IHasTouchstones {
+	currentForm: Form;
+	// baseFormMods!: { [key: string]: FormMods };
 
-// 	player?: string;
-// 	chronicle?: string;
+	kuruthTriggers: { passive: string; common: string; specific: string };
 
-// 	virtueAnchor!: string;
-// 	viceAnchor!: string;
+	huntersAspect: string;
 
-// 	concept?: string;
+	moonGift1: Ability;
+	moonGift2: Ability;
 
-// 	splat!: EnumSplat;
-// 	splatObj!: Splat;
+	moonGifts: { [key: string]: Ability };
 
-// 	subType!: Ref<string>; // Clan/Auspice/Path
-// 	subTypeObj!: ComputedRef<SubType | undefined>; // Clan/Auspice/Path
+	shadowGifts: string[];
+	wolfGifts: string[];
 
-// 	faction?: string; // Cabal/Coterie/Pack
+	rites: string[];
+}
 
-// 	organization!: Ref<string>; // Order/Covenant/Tribe
-// 	organizationObj!: ComputedRef<Organization | undefined>;
+export interface FormMods {
+	intelligenceMod: number;
+	witsMod: number;
+	resolveMod: number;
 
+	strengthMod: number;
+	dexterityMod: number;
+	staminaMod: number;
 
+	presenceMod: number;
+	manipulationMod: number;
+	composureMod: number;
 
+	sizeMod: number;
+	speedMod: number;
 
-// 	legacy?: string; // Legacy/Bloodline/Lodge
+	perceptionMod: number;
 
-// 	baseAttributes!: Attributes;
-// 	attributes: ComputedRef<Attributes>;
+	defenseCalcMax?: boolean;
+	defenseMod?: number;
 
-// 	skills: { [index: string]: number } = {};
-// 	specialties: { [index: string]: string[] } = {};
+	armorMod: Armor;
+}
 
-// 	// abilityArr: Ability[] | Ref<Ability[]>;
-// 	abilities: { [index: string]: Ability };
+export interface Form extends FormMods {
+	name: string;
+	desc: string;
+	traits: string[];
 
-// 	merits: { [key: string]: Ability | Merit } = {};
+	[index: string]: any;
+}
 
-// 	// abilities: {[index: string]: Ability} = {};;
-// 	// merits: {[index: string]: Ability} = {};
 
+export class WerewolfCharacter extends SupernaturalCharacter implements IWerewolfCharacter {
 
-// 	healthTrack: Ref<number[]> = ref([]);
-// 	woundPenalty: ComputedRef<number>;
-// 	maxHealth: ComputedRef<number>;
+	// currentForm: Form;
+	get currentForm(): Form {
+		return this.forms[this.data.get("currentForm") as string] || {};
+	}
 
-// 	willpower?: number;
-// 	spentWillpowerDots = 0;
-// 	// maxWillpower?: number;
+	get forms(): { [key: string]: Form } {
+		return (SPLATS[EnumSplat.WEREWOLF] as any).forms as { [key: string]: Form };
+	}
 
-// 	power = ref(1); // Gnosis/Primal Urge/Blood Potency
-
-// 	fuel = 0;
-// 	maxFuel: ComputedRef<number>;
-
-// 	integrityTrait: RefType<number> = 7;
-// 	integrityTrack?: number[];
-
-// 	touchstones: { name: string; type?: string }[] = [];
-
-// 	conditions: string[] = [];
-
-// 	aspirations: string[] = [];
-
-// 	baseSize = 5;
-// 	size: ComputedRef<number>;
-
-// 	defenseCalcMax: ComputedRef<boolean>;
-// 	defense: ComputedRef<number>;
-
-// 	speed: ComputedRef<number>;
-
-// 	armor: ComputedRef<Armor>;
-// 	baseArmor: Armor = {
-// 		general: 0,
-// 		ballistic: 0
-// 	};
-
-// 	weapons: Weapon[] = [
-// 		{
-// 			name: "",
-// 			damage: "",
-// 			range: "",
-// 			clip: "",
-// 			initative: 0,
-// 			strength: 0,
-// 			size: 0,
-// 		}
-// 	];
-
-// 	// initative?: number;
-
-// 	beats?= 0;
-// 	experience?= 0;
-
-// 	alternateBeats?= 0;
-// 	alternateExperience?= 0;
-
-// 	traitMods: ComputedRef<TraitMod[]> = computed(() => {
-// 		const meritMods = Object.keys(this.merits)
-// 			.map((key) => this.merits[key])
-// 			.filter(el => (el as any).getOptions)
-// 			.flatMap((el) => (el as Merit).getTraitMods(this));
-
-// 		return [...this.getTraitMods(), ...meritMods];
-// 	});
-
-// 	getNum(arg: string) {
-// 		return getNum.call(this, arg);
-// 	}
-
-// 	constructor(opts: Character) {
-// 		this.splatObj = SPLATS[opts.splat];
-
-// 		if (this.splatObj.integrityTrackType === "healthTrack") {
-// 			this.integrityTrack = [];
-// 		}
-
-// 		{
-// 			const defaultAbl: { [index: string]: Ability } = {};
-// 			const tt = this.splatObj.abilities || {};
-
-// 			Object.keys(tt).forEach((key) => {
-// 				defaultAbl[key] = { name: computed(() => tt[key]), level: 0 };
-// 			});
-// 			Object.assign(this, opts);
-
-// 			// this.maxHealth
-// 			// this.integrityTrait = ref(this.integrityTrait);
-
-
-// 			// const arr: string[] = [];
-
-// 			// watch(
-// 			// 	() => Object.keys(this.merits),
-// 			// 	(keys, prevKeys) => {
-// 			// 		console.log("merits");
-
-// 			// 		const newKeys = keys.filter(x => prevKeys.includes(x));
-
-// 			// 		newKeys.forEach(key => {
-// 			// 			const val = this.merits[key];
-
-// 			// 			const m = MERITS[key];
-
-// 			// 			if (!(val as Merit).getOptions && m) {
-// 			// 				const meritObj = new m(this, val);
-
-// 			// 				// arr.push(key);
-
-// 			// 				this.merits[key] = meritObj;
-// 			// 			}
-// 			// 		});
-
-// 			// 		// // Object.values(val).forEach(value => {
-// 			// 		// 	let key = nameToKey(value.name);
-
-// 			// 		// 	if (key.includes("(")) {
-// 			// 		// 		key = key.substr(0, key.indexOf("(")-1);
-// 			// 		// 	}
-
-// 			// 		// 	const m = MERITS[key];
-
-// 			// 		// 	// if ((value as Merit).getOptions) {
-// 			// 		// 	// 	console.log("replace", value.name, val, prevVal);
-// 			// 		// 	// }
-
-// 			// 		// 	if (!arr.includes(key) && !(value as any).getOptions && m) {
-// 			// 		// 		const meritObj = new m(this, value);
-// 			// 		// 		// Object.assign(m, value);
-
-// 			// 		// 		arr.push(key);
-
-// 			// 		// 		this.merits[key] = meritObj;
-// 			// 		// 		console.log(this.merits[key]);
-// 			// 		// 	}
-// 			// 		// });
-// 			// 	},
-// 			// 	{ deep: true }
-// 			// );
-
-// 			const ablTemp = Object.assign({}, defaultAbl, (this as any).abilities);
-
-// 			const custom: any = {};
-
-// 			Object.keys(ablTemp).forEach(key => {
-// 				const value = ablTemp[key];
-// 				const def = defaultAbl[key];
-
-// 				if (!def) {
-// 					custom[key] = value;
-// 					delete ablTemp[key];
-// 				} else {
-// 					value.name = def.name;
-// 				}
-// 			});
-
-// 			// ablTemp = sortObj(ablTemp);
-
-// 			Object.keys(custom)
-// 				.forEach(key => ablTemp[key] = custom[key]);
-
-// 			this.abilities = reactive(ablTemp);
-// 		}
-
-// 		this.power = ref(this.power || 1);
-
-// 		this.merits = reactive(this.merits);
-// 		this.baseAttributes = reactive(this.baseAttributes || {
-// 			intelligence: 1,
-// 			wits: 1,
-// 			resolve: 1,
-// 			strength: 1,
-// 			dexterity: 1,
-// 			stamina: 1,
-// 			presence: 1,
-// 			manipulation: 1,
-// 			composure: 1
-// 		});
-
-// 		this.subType = ref(this.subType || "");
-// 		this.organization = ref(this.organization || "");
-
-// 		this.healthTrack = ref(this.healthTrack as any);
-
-// 		Object.keys(this.merits).forEach((key) => {
-// 			const value = this.merits[key];
-
-// 			const meritKey = key.includes("(") ? key.substr(0, key.indexOf("(") - 1) : key;
-
-// 			const m = MERITS[meritKey];
-
-// 			if (!(value as Merit).getOptions && m) {
-// 				this.merits[key] = new m(this, value);
-// 			}
-// 		});
-
-// 		this.subTypeObj = computed(() => {
-// 			return Object.values(this.splatObj.subTypes).find(el =>
-// 				nameToKey(unref(el.name)) === nameToKey(unref(this.subType))
-// 			);
-// 		});
-
-// 		this.organizationObj = computed(() => {
-// 			return (Object.entries(this.splatObj.organizations)
-// 				.find(el =>
-// 					el[0] === unref(this.organization)) || [])[1];
-// 		});
-
-// 		this.attributes = computed(() => {
-// 			return {
-// 				intelligence: this.baseAttributes.intelligence + this.mod("intelligence"),
-// 				wits: this.baseAttributes.wits + this.mod("wits"),
-// 				resolve: this.baseAttributes.resolve + this.mod("resolve"),
-
-// 				strength: this.baseAttributes.strength + this.mod("strength"),
-// 				dexterity: this.baseAttributes.dexterity + this.mod("dexterity"),
-// 				stamina: this.baseAttributes.stamina + this.mod("stamina"),//((self.abilities.value.resilience || {}).level || 0),//vue.getNum("this.character.abilities.value.resilience.level"),
-
-// 				presence: this.baseAttributes.presence + this.mod("presence"),
-// 				manipulation: this.baseAttributes.manipulation + this.mod("manipulation"),
-// 				composure: this.baseAttributes.composure + this.mod("composure"),
-// 			};
-// 		});
-
-// 		this.size = computed({
-// 			get: () => {
-// 				// console.log(this.baseSize, this.mod("size"));
-// 				return this.baseSize + this.mod("size");
-// 			},
-// 			set: (val) => {
-// 				this.baseSize = val - this.mod("size");
-// 			}
-// 		});
-
-// 		this.maxHealth = computed(() => {
-// 			return this.attributes.value.stamina + this.size.value + this.mod("health");
-// 		});
-
-// 		this.woundPenalty = computed(() => {
-// 			return Math.min((this.healthTrack.value[this.maxHealth.value - 1] !== 0 ?
-// 				-3 : this.healthTrack.value[this.maxHealth.value - 2] !== 0 ?
-// 					-2 : this.healthTrack.value[this.maxHealth.value - 3] !== 0 ?
-// 						-1 : 0) + this.mod("woundPenalty"), 0);
-// 		});
-
-// 		this.speed = computed(() => {
-// 			return this.attributes.value.strength + this.attributes.value.dexterity + 5 + this.mod("speed");
-// 		});
-
-// 		this.defenseCalcMax = computed(() => {
-// 			return !!this.mod("defenseCalcMax");
-// 		});
-// 		this.defense = computed(() => {
-// 			return (this.defenseCalcMax ? Math.max : Math.min)(this.attributes.value.dexterity, this.attributes.value.wits) + def(() => this.skills.athletics).value + this.mod("defense");
-// 		});
-
-// 		this.armor = computed(() => {
-// 			return {
-// 				general: this.baseArmor.general + this.mod("generalArmor"),
-// 				ballistic: this.baseArmor.ballistic + this.mod("ballisticArmor")
-// 			};
-// 		});
-
-// 		this.maxFuel = computed(() => {
-// 			const power = unref(this.power);
-// 			return power >= 5 ?
-// 				power >= 9 ?
-// 					power === 10 ?
-// 						75 :
-// 						50
-// 					: 10 + (power - 4) * 5
-// 				: 10 + power - 1;
-// 		});
-// 	}
-
-// 	getTraitMods(): TraitMod[] {
-// 		return [];
-// 	}
-
-// 	mod(traitName: string): number {
-// 		const na = traitName.toLowerCase();
-
-// 		return unref(this.traitMods)
-// 			.filter(el => el.trait && el.trait.toLowerCase() === na)
-// 			.map(el => el.mod)
-// 			.reduce((prev: any, val: any) => prev + unref(val), 0) || 0;
-// 	}
-
-// 	getData() {
-// 		const raw = toRaw(this) as any;
-// 		const obj = { ...this } as any;
-
-// 		delete obj.traitMods;
-// 		delete obj.splatObj;
-// 		delete obj.size;
-
-// 		Object.keys(obj).forEach(key => {
-// 			if (isRef(raw[key]) && isReadonly(raw[key])) {
-// 				delete obj[key];
-// 			}
-// 		});
-
-// 		return obj;
-// 	}
-// }
-
-// export class Rote {
-// 	arcanum!: string;
-// 	level!: number;
-// 	spell!: string;
-// 	creator?: string;
-// 	roteSkill!: string;
-
-// 	constructor(opts: Rote) {
-// 		Object.assign(this, {
-// 			arcanum: "",
-// 			level: 0,
-// 			spell: "",
-// 			creator: "",
-// 			roteSkill: ""
-// 		}, opts);
-// 	}
-// }
-
-// export class MageCharacter extends Character {
-// 	obsessions: string[];
-
-// 	activeSpells!: string[];
-// 	yantras!: string[];
-// 	magicalTools!: string[];
-// 	praxes!: string[];
-// 	inuredSpells!: string[];
-
-// 	nimbus: string[];
-
-// 	arcanaAttainments!: string[];
-// 	legacyAttainments!: string[];
-
-// 	baseRoteSkills!: Ref<string[]>;
-// 	roteSkills!: ComputedRef<string[]>;
-
-// 	rotes!: Rote[];
-
-// 	constructor(opts: MageCharacter) {
-// 		super(opts as any);
-
-// 		this.obsessions = opts.obsessions || [];
-
-// 		this.activeSpells = opts.activeSpells || [];
-// 		this.yantras = opts.yantras || [];
-// 		this.magicalTools = opts.magicalTools || [];
-// 		this.praxes = opts.praxes || [];
-// 		this.inuredSpells = opts.inuredSpells || [];
-
-// 		this.arcanaAttainments = opts.arcanaAttainments || [];
-// 		this.legacyAttainments = opts.legacyAttainments || [];
-
-// 		this.baseRoteSkills = ref(opts.baseRoteSkills || []);
-
-// 		this.nimbus = opts.nimbus || [];
-
-// 		this.roteSkills = computed({
-// 			get: () => {
-// 				// const order = SPLATS[EnumSplat.MAGE].organizations[unref(this.organization)];
-// 				const order = this.organizationObj.value;
-
-// 				console.log(SPLATS[EnumSplat.MAGE].organizations);
-
-// 				if (order) {
-// 					return order.skills || [];
-// 				}
-// 				return unref(this.baseRoteSkills);
-// 			},
-// 			set: (val) => {
-// 				const order = this.organizationObj.value;
-
-// 				if (!order) {
-// 					this.baseRoteSkills.value = val;
-// 				}
-// 			}
-// 		});
-
-// 		this.rotes = opts.rotes || [];
-// 	}
-// }
-
-// export interface Devotion {
-// 	name: string;
-// 	cost: string;
-// 	disciplines: string[];
-// 	dicePool: string;
-// 	book: string;
-// 	pageNr: number;
-// }
-
-// export class VampireCharacter extends Character {
-
-// 	devotions: Devotion[];
-
-// 	constructor(opts: VampireCharacter) {
-// 		super(opts as any);
-
-// 		this.devotions = opts.devotions || [];
-
-// 		const maxFuel = this.maxFuel;
-
-// 		this.maxFuel = computed(() => this.power.value === 0 ? this.attributes.value.stamina : unref(maxFuel));
-// 	}
-
-// 	getTraitMods() {
-// 		return super.getTraitMods().concat([
-// 			{ trait: "stamina", mod: def(() => this.abilities.resilience.level) },
-// 			{ trait: "strength", mod: def(() => this.abilities.vigor.level) },
-
-// 			{ trait: "defense", mod: def(() => this.abilities.celerity.level) }
-// 		]);
-// 	}
-// }
-
-// export class WerewolfCharacter extends Character {
-
-// 	currentForm: Ref<string> = ref("hishu");
-// 	baseFormMods!: { [key: string]: FormMods };
-
-// 	kuruthTriggers?: { passive: string; common: string; specific: string };
-
-// 	huntersAspect: string;
-
-// 	// moonGifts: [Ability | undefined, Ability | undefined];
-
-// 	moonGift1: ComputedRef<Ability>;
-// 	moonGift2: Ref<Ability>;
-
-// 	moonGifts: ComputedRef<{ [key: string]: Ability }>;
-
-// 	shadowGifts: string[];
-// 	wolfGifts: string[];
-
-// 	rites: string[];
-
-// 	constructor(opts: WerewolfCharacter) {
-// 		super(opts);
-
-// 		this.baseFormMods = opts.baseFormMods || {};
-// 		this.kuruthTriggers = Object.assign({
-// 			passive: "",
-// 			common: "",
-// 			specific: ""
-// 		}, this.kuruthTriggers);
-
-// 		this.huntersAspect = opts.huntersAspect || "";
-
-// 		// this.moonGifts = opts.moonGifts || [
-// 		// 	{name: "", level: 0},
-// 		// 	{name: "", level: 0}
-// 		// ];
-
-// 		// this.moonGifts = opts.moonGifts || {};
-
-// 		// const auspice = computed(() => {
-// 		// 	const key = Object.keys(this.splatObj.subTypes)
-// 		// 		.find(el => el === nameToKey(this.subType.value)) || "";
-// 		// 	return this.splatObj.subTypes[key];
-// 		// });
-
-// 		const woundPenalty = this.woundPenalty;
-
-// 		this.woundPenalty = computed(() => {
-// 			return unref(this.currentForm) === "gauru" ? 0 : unref(woundPenalty);
-// 		});
-
-// 		this.moonGift1 = computed(() => {
-// 			if (this.subTypeObj.value) {
-// 				const key = (this.subTypeObj.value.moonGifts as string[])[0];
-// 				const renown = this.subTypeObj.value.abilities[0];
-
-// 				return {
-// 					name: t("splat.werewolf.gift.moon." + key),
-// 					key,
-// 					level: this.abilities[renown].level
-// 				} as Ability;
-// 			}
-// 			return unref(opts.moonGift1) || { name: "", key: "", level: 0 } as Ability;
-// 		});
-
-// 		this.moonGift2 = ref(opts.moonGift2 || { name: "", level: 0 });
-// 		this.moonGifts = computed(() => {
-// 			return {
-// 				[this.moonGift1.value.key || nameToKey(unref(this.moonGift1.value.name))]: unref(this.moonGift1),
-// 				[this.moonGift2.value.key || nameToKey(unref(this.moonGift2.value.name))]: unref(this.moonGift2),
-// 			};
-// 		});
-
-// 		this.shadowGifts = opts.shadowGifts || [];
-// 		this.wolfGifts = opts.wolfGifts || [];
-// 		this.rites = opts.rites || [];
-// 	}
-
-// 	getTraitMods() {
-// 		return super.getTraitMods().concat([
-// 			{
-// 				trait: "health", mod: def(() => {
-// 					const full = this.moonGifts.value.full;
-// 					if (full) {
-// 						// const purity = def(() => this.abilities.purity.level, 0).value;
-// 						return full.level >= 2 ? full.level : 0;
-// 					}
-// 					return 0;
-// 				})
-// 			},
-// 			...ATTRIBUTES.flat().map(attr => {
-// 				return { trait: attr, mod: def(() => (this.currentFormObj().value as any)[attr + "Mod"]) };
-// 			}),
-// 			{ trait: "size", mod: def(() => this.currentFormObj().value.sizeMod) },
-// 			{ trait: "speed", mod: def(() => this.currentFormObj().value.speedMod) },
-// 			{ trait: "perception", mod: def(() => this.currentFormObj().value.perceptionMod) },
-
-// 			{ trait: "ballisticArmor", mod: def(() => this.currentFormObj().value.armorMod.ballistic || 0) },
-// 			{ trait: "defenseCalcMax", mod: def(() => this.currentFormObj().value.defenseCalcMax) },
-
-// 			{ trait: "generalArmor", mod: def(() => this.currentFormObj().value.armorMod.general || 0) },
-// 			{ trait: "ballisticArmor", mod: def(() => this.currentFormObj().value.armorMod.ballistic || 0) },
-// 		]);
-// 	}
-
-// 	currentFormObj(): WritableComputedRef<Form> {
-// 		return this.getForm(this.currentForm);
-// 	}
-
-// 	getForm(name: Ref<string> | string): WritableComputedRef<Form> {
-// 		return computed({
-// 			get: () => {
-// 				const key = isRef(name) ? name.value : name;
-
-// 				const form = WEREWOLF_FORMS[key] || {
-// 					name: "",
-// 					desc: "",
-
-// 					intelligenceMod: 0,
-// 					witsMod: 0,
-// 					resolveMod: 0,
-
-// 					strengthMod: 0,
-// 					staminaMod: 0,
-// 					dexterityMod: 0,
-
-// 					presenceMod: 0,
-// 					manipulationMod: 0,
-// 					composureMod: 0,
-
-// 					sizeMod: 0,
-// 					speedMod: 0,
-// 					perceptionMod: 0,
-
-// 					traits: []
-// 				};
-
-// 				const baseMod: FormMods = Object.assign({
-// 					intelligenceMod: 0,
-// 					witsMod: 0,
-// 					resolveMod: 0,
-
-// 					strengthMod: 0,
-// 					staminaMod: 0,
-// 					dexterityMod: 0,
-
-// 					presenceMod: 0,
-// 					manipulationMod: 0,
-// 					composureMod: 0,
-
-// 					sizeMod: 0,
-// 					speedMod: 0,
-// 					perceptionMod: 0,
-
-// 					armorMod: {}
-// 				}, this.baseFormMods[key]);
-
-// 				// self.baseFormMods[key] = baseMod;
-
-// 				return Object.assign({}, form, {
-// 					intelligenceMod: form.intelligenceMod + baseMod.intelligenceMod + this.mod(key + "IntelligenceMod"),
-// 					witsMod: form.witsMod + baseMod.witsMod + this.mod(key + "WitsMod"),
-// 					resolveMod: form.resolveMod + baseMod.resolveMod + this.mod(key + "ResolveMod"),
-
-// 					strengthMod: form.strengthMod + baseMod.strengthMod + this.mod(key + "StrengthMod"),
-// 					dexterityMod: form.dexterityMod + baseMod.dexterityMod + this.mod(key + "DexterityMod"),
-// 					staminaMod: form.staminaMod + baseMod.staminaMod + this.mod(key + "StaminaMod"),
-
-// 					presenceMod: form.presenceMod + baseMod.presenceMod + this.mod(key + "PresenceMod"),
-// 					manipulationMod: form.manipulationMod + baseMod.manipulationMod + this.mod(key + "ManipulationMod"),
-// 					composureMod: form.composureMod + baseMod.composureMod + this.mod(key + "ComposureMod"),
-
-// 					sizeMod: form.sizeMod + baseMod.sizeMod + this.mod(key + "SizeMod"),
-// 					speedMod: form.speedMod + baseMod.speedMod + this.mod(key + "SpeedMod"),
-// 					perceptionMod: form.perceptionMod + baseMod.perceptionMod + this.mod(key + "PerceptionMod"),
-
-// 					defenseCalcMax: !!this.mod(key + "DefenseCalcMax") as boolean,
-// 					defenseMod: this.mod(key + "DefenseMod"),
-
-// 					armorMod: {
-// 						general: (form.armorMod.general || 0) + (baseMod.armorMod.general || 0) + this.mod(key + "generalArmorMod"),
-// 						ballistic: (form.armorMod.ballistic || 0) + (baseMod.armorMod.ballistic || 0) + this.mod(key + "ballisticArmorMod")
-// 					},
-
-// 					// TODO
-// 					traits: form.traits.map((el: string, i: number) => {
-// 						return i === 0 && el.includes("+") ? "" : el;
-// 					})
-// 				});
-// 			},
-// 			set: (val) => {
-// 				const key = isRef(name) ? name.value : name;
-
-// 				const form = WEREWOLF_FORMS[key] || {
-// 					name: "",
-// 					desc: "",
-
-// 					strengthMod: 0,
-// 					staminaMod: 0,
-// 					dexterityMod: 0,
-// 					manipulationMod: 0,
-
-// 					sizeMod: 0,
-// 					speedMod: 0,
-// 					perceptionMod: 0,
-
-// 					traits: []
-// 				};
-
-// 				val = Object.assign({}, form, val);
-
-// 				this.baseFormMods[key] = Object.assign({}, {
-// 					intelligenceMod: val.intelligenceMod - form.intelligenceMod - this.mod(key + "IntelligenceMod"),
-// 					witsMod: val.witsMod - form.witsMod - this.mod(key + "WitsMod"),
-// 					resolveMod: val.resolveMod - form.resolveMod - this.mod(key + "ResolveMod"),
-
-// 					strengthMod: val.strengthMod - form.strengthMod - this.mod(key + "StrengthMod"),
-// 					staminaMod: val.staminaMod - form.staminaMod - this.mod(key + "DexterityMod"),
-// 					dexterityMod: val.dexterityMod - form.dexterityMod - this.mod(key + "StaminaMod"),
-
-// 					presenceMod: val.presenceMod - form.presenceMod - this.mod(key + "PresenceMod"),
-// 					manipulationMod: val.manipulationMod - form.manipulationMod - this.mod(key + "ManipulationMod"),
-// 					composureMod: val.composureMod - form.composureMod - this.mod(key + "ComposureMod"),
-
-// 					sizeMod: val.sizeMod - form.sizeMod - this.mod(key + "SizeMod"),
-// 					speedMod: val.speedMod - form.speedMod - this.mod(key + "SpeedMod"),
-// 					perceptionMod: val.perceptionMod - form.perceptionMod - this.mod(key + "PerceptionMod"),
-
-// 					armorMod: {
-// 						general: (val.armorMod.general || 0) - (form.armorMod.general || 0) - this.mod(key + "generalArmorMod"),
-// 						ballistic: (val.armorMod.ballistic || 0) - (form.armorMod.ballistic || 0) - this.mod(key + "ballisticArmorMod"),
-// 					}
-// 				});
-// 			}
-// 		});
-// 	}
-
-// 	getForms(): { [key: string]: Form } {
-// 		const forms: { [key: string]: Form } = {};
-
-// 		Object.keys(WEREWOLF_FORMS)
-// 			.forEach(el => forms[el] = this.getForm(el) as unknown as Form);
-
-// 		return reactive(forms);
-// 	}
-// }
-
-// export class ChangelingCharacter extends Character {
-// 	constructor(opts: ChangelingCharacter) {
-// 		super(opts as any);
-
-// 		this.integrityTrait = computed(() => this.attributes.value.wits + this.attributes.value.composure);
-// 	}
-// }
-
-// export interface EphemeralAttributes extends Attributes {
+	get size(): number {
+		return super.size + this.currentForm.sizeMod;
+	}
+
+	set size(val: number) {
+		super.size = val - this.currentForm.sizeMod;
+	}
+
+	get speed(): number {
+		return super.speed + this.currentForm.speedMod;
+	}
+
+	get maxHealth(): number {
+		return super.maxHealth + (this.moonGifts.full.level >= 2 ?
+			this.abilities.purity.level : 0);
+	}
+
+	kuruthTriggers: { passive: string; common: string; specific: string; };
+	huntersAspect: string;
+
+	// moonGift1: Ability;
+
+	get moonGift1(): Ability {
+		if (this.subType.name) {
+			const key = (this.subType.moonGifts as string[])[0];
+			const renown = this.subType.abilities[0];
+
+			return {
+				// name: "splat.werewolf.gift.moon." + key,
+				key,
+				level: this.abilities[renown].level
+			} as Ability;
+		}
+		return this.data.get("moonGift1") as Ability || { name: "", key: "", level: 0 } as Ability;
+	}
+	set moonGift1(val: Ability) {
+		if (!this.subType.name) {
+			this.data.set("moonGift1", val);
+		}
+	}
+
+	moonGift2: Ability;
+
+	get moonGifts(): { [key: string]: Ability } {
+		if (!this.moonGift2.key) this.moonGift2.key = "NEW";
+		return {
+			[this.moonGift1.key || ""]: this.moonGift1,
+			[this.moonGift2.key || ""]: this.moonGift2,
+		};
+	}
+
+	shadowGifts: string[];
+	wolfGifts: string[];
+	rites: string[];
+
+	touchstones: string[];
+
+	get perception(): number {
+		return super.perception + this.currentForm.perceptionMod;
+	}
+
+	constructor(opts: IWerewolfCharacter) {
+		super(opts);
+		this.touchstones = opts.touchstones || [];
+
+		this.data.set("currentForm", opts.currentForm || "hishu");
+
+		this.kuruthTriggers = opts.kuruthTriggers || {
+			passive: "",
+			common: "",
+			specific: ""
+		};
+		this.huntersAspect = opts.huntersAspect || "";
+
+		this.moonGift2 = opts.moonGift2 || { name: "", level: 0 };
+
+		this.shadowGifts = opts.shadowGifts || [];
+		this.wolfGifts = opts.wolfGifts || [];
+		this.rites = opts.rites || [];
+
+		const descs = Object.getOwnPropertyDescriptors(this.attributes);
+
+		const _defProp = (key: string, func: () => number) => ({
+			get: () => ((descs[key].get || (() => 0))() || 1) + func(),
+			set: (val: number) => (descs[key].set || (() => 0))(val - func()),
+			configurable: true
+		});
+		this.attributes = Object.defineProperties(this.attributes, {
+			intelligence: _defProp("intelligence", () => this.currentForm.intelligenceMod),
+			wits: _defProp("wits", () => this.currentForm.witsMod),
+			resolve: _defProp("resolve", () => this.currentForm.resolveMod),
+
+			strength: _defProp("strength", () => this.currentForm.strengthMod),
+			dexterity: _defProp("dexterity", () => this.currentForm.dexterityMod),
+			stamina: _defProp("stamina", () => this.currentForm.staminaMod),
+
+			presence: _defProp("presence", () => this.currentForm.presenceMod),
+			manipulation: _defProp("manipulation", () => this.currentForm.manipulationMod),
+			composure: _defProp("composure", () => this.currentForm.composureMod),
+		});
+	}
+}
+// interface ISupernaturalCharacter {
 // 	power: number;
-// 	finesse: number;
-// 	resistance: number;
+
+// 	fuel: number;
+// 	maxFuel: number;
+
+// 	abilities: { [index: string]: Ability }
 // }
 
-// export class EphemeralCharacter extends Character {
-
-// 	baseAttributes: EphemeralAttributes;
-// 	attributes: ComputedRef<EphemeralAttributes>;
-
-// 	numina: string[];
-
-// 	ban: string;
-// 	bane: string;
-
-// 	constructor(opts: EphemeralCharacter) {
-// 		super(opts as any);
-
-// 		// this.power = opts.power || 0;
-// 		// this.finesse = opts.finesse || 0;
-// 		// this.resistance = opts.finesse || 0;
-
-// 		this.numina = opts.numina || [];
-
-// 		this.ban = opts.ban || "";
-// 		this.bane = opts.bane || "";
-
-// 		this.maxFuel = computed(() => this.power.value === 5 ? 50 : 5 + (5 * this.power.value));
-// 		this.baseAttributes = reactive(opts.baseAttributes || {
-// 			power: 1,
-// 			finesse: 1,
-// 			resistance: 1
-// 		});
-
-// 		this.attributes = computed(() => {
-// 			const power = this.baseAttributes.power + this.mod("power");
-// 			const finesse = this.baseAttributes.finesse + this.mod("finesse");
-// 			const resistance = this.baseAttributes.resistance + this.mod("resistance");
-// 			return {
-// 				power,
-// 				finesse,
-// 				resistance,
-
-// 				intelligence: power,
-// 				strength: power,
-// 				presence: power,
-
-// 				wits: finesse,
-// 				dexterity: finesse,
-// 				manipulation: finesse,
-
-// 				resolve: resistance,
-// 				stamina: resistance,
-// 				composure: resistance
-// 			};
-// 		});
-// 	}
+// interface IHasIntegrity {
+// 	integrityTrait: number;
 // }
 
-// export function createCharacter<T extends Character>(opts: T): T {
-// 	return new SPLATS[opts.splat].characterFactory(opts) as T;
+// interface IHasIntegrityTrack extends IHasIntegrity {
+// 	integrityTrack?: number[];
 // }
+
+// interface IHasTouchstones {
+// 	touchstones: { name: string; type?: string }[];
+// }
+
+// interface IEarnsBeats {
+// 	beats: number;
+// 	experience: number;
+// }
+
+// interface IPC extends IHasIntegrity, IEarnsBeats {
+
+// }
+
+export function createCharacter<T extends Character>(opts: T): T {
+	try {
+		const splat = SPLATS[opts.splat as unknown as EnumSplat] || SPLATS[EnumSplat.MORTAL];
+		return reactive(new splat.characterFactory(opts) as any) as T;
+	} catch (e) {
+		console.error(e);
+	}
+}
