+++ src/definitions/index.ts
@@ -1,3 +1,3 @@
-export * from "./NewCharacter";
-export * from "./NewSplat";
+export * from "./Character";
+export * from "./Splat";
 export * from "./Merit";
